feat: Add future messaging
* Improve storage integration
* Update submodule shared-core
* Modify business logic

+++ Conduit.Person.BusinessLogic/FollowingsManager.cs
@@ -1,35 +1,66 @@
 ﻿using System.Net;
+using Conduit.Shared.Events.Models.Profiles.CreateFollowing;
+using Conduit.Shared.Events.Models.Profiles.RemoveFollowing;
+using Conduit.Shared.Events.Services;
 
 namespace Conduit.Person.BusinessLogic;
 
 public class FollowingsManager : IFollowingsManager
 {
+    private readonly IEventProducer<CreateFollowingEventModel>
+        _createFollowingProducer;
+
     private readonly IProfileRepository _profileRepository;
 
+    private readonly IEventProducer<RemoveFollowingEventModel>
+        _removeFollowingProducer;
+
     public FollowingsManager(
-        IProfileRepository profileRepository)
+        IProfileRepository profileRepository,
+        IEventProducer<CreateFollowingEventModel> createFollowingProducer,
+        IEventProducer<RemoveFollowingEventModel> removeFollowingProducer)
     {
         _profileRepository = profileRepository;
+        _createFollowingProducer = createFollowingProducer;
+        _removeFollowingProducer = removeFollowingProducer;
     }
 
     public async Task<ExecutionResult<ProfileResponse>> AssignFollowingAsync(
         FollowingInfo followingInfo,
         CancellationToken cancellationToken = default)
     {
-        var profileResponse =
+        var (profile, _) = await _profileRepository.FindAsync(followingInfo);
+        if (profile is null || profile.Profile.Following)
+        {
+            return GetResult(null);
+        }
+
+        var (newFollowedProfile, followedId) =
             await _profileRepository.AddFollowingAsync(followingInfo);
 
-        return GetResult(profileResponse);
+        await _createFollowingProducer.ProduceEventAsync(
+            new(followingInfo.FollowerUserId, followedId!.Value));
+
+        return GetResult(newFollowedProfile);
     }
 
     public async Task<ExecutionResult<ProfileResponse>> DeAssignFollowingAsync(
         FollowingInfo followingInfo,
         CancellationToken cancellationToken = default)
     {
-        var profileResponse =
+        var (profile, _) = await _profileRepository.FindAsync(followingInfo);
+        if (profile is null || profile.Profile.Following == false)
+        {
+            return GetResult(null);
+        }
+
+        var (newFollowedProfile, followedId) =
             await _profileRepository.RemoveFollowingAsync(followingInfo);
 
-        return GetResult(profileResponse);
+        await _removeFollowingProducer.ProduceEventAsync(
+            new(followingInfo.FollowerUserId, followedId!.Value));
+
+        return GetResult(newFollowedProfile);
     }
 
     private static ExecutionResult<ProfileResponse> GetResult(
